+++ curator/indexlist.py
@@ -100,6 +100,7 @@ class IndexList(object):
             'kibana': self.filter_kibana,
             'none': self.filter_none,
             'opened': self.filter_opened,
+            'period': self.filter_period,
             'pattern': self.filter_by_regex,
             'space': self.filter_by_space,
         }
@@ -432,11 +433,12 @@ class IndexList(object):
         )
         for index in self.working_list():
             try:
+                age = int(self.index_info[index]['age'][self.age_keyfield])
                 msg = (
                     'Index "{0}" age ({1}), direction: "{2}", point of '
                     'reference, ({3})'.format(
                         index,
-                        int(self.index_info[index]['age'][self.age_keyfield]),
+                        age,
                         direction,
                         PoR
                     )
@@ -444,9 +446,9 @@ class IndexList(object):
                 # Because time adds to epoch, smaller numbers are actually older
                 # timestamps.
                 if direction == 'older':
-                    agetest = self.index_info[index]['age'][self.age_keyfield] < PoR
+                    agetest = age < PoR
                 else:
-                    agetest = self.index_info[index]['age'][self.age_keyfield] > PoR
+                    agetest = age > PoR
                 self.__excludify(agetest, exclude, index, msg)
             except KeyError:
                 self.loggit.debug(
@@ -795,6 +797,73 @@ class IndexList(object):
             condition = True if idx <= count else False
             self.__excludify(condition, exclude, index, msg)
             idx += 1
+
+    def filter_period(
+        self, source='name', range_from=None, range_to=None, timestring=None,
+        unit=None, field=None, stats_result='min_value', 
+        week_starts_on='sunday', epoch=None, exclude=False,
+        ):
+        """
+        Match `indices` within ages within a given period.
+
+        :arg source: Source of index age. Can be one of 'name', 'creation_date',
+            or 'field_stats'
+        :arg range_from: How many ``unit``s in the past/future is the origin?
+        :arg range_to: How many ``unit``s in the past/future is the end point?
+        :arg timestring: An strftime string to match the datestamp in an index
+            name. Only used for index filtering by ``name``.
+        :arg unit: One of ``hours``, ``days``, ``weeks``, ``months``, or 
+            ``years``.
+        :arg unit_count: The number of ``unit``s. ``unit_count`` * ``unit`` will
+            be calculated out to the relative number of seconds.
+        :arg field: A timestamp field name.  Only used for ``field_stats`` based
+            calculations.
+        :arg stats_result: Either `min_value` or `max_value`.  Only used in
+            conjunction with `source`=``field_stats`` to choose whether to
+            reference the minimum or maximum result value.
+        :arg week_starts_on: Either ``sunday`` or ``monday``. Default is 
+            ``sunday``
+        :arg epoch: An epoch timestamp used to establish a point of reference 
+            for calculations. If not provided, the current time will be used.
+        :arg exclude: If `exclude` is `True`, this filter will remove matching
+            indices from `indices`. If `exclude` is `False`, then only matching
+            indices will be kept in `indices`.
+            Default is `False`
+        """
+
+        self.loggit.debug('Filtering indices by age')
+        try:
+            start, end = date_range(
+                unit, range_from, range_to, epoch, week_starts_on=week_starts_on
+            )
+        except Exception as e:
+            report_failure(e)
+
+        self._calculate_ages(
+            source=source, timestring=timestring, field=field,
+            stats_result=stats_result
+        )
+        for index in self.working_list():
+            try:
+                age = int(self.index_info[index]['age'][self.age_keyfield])
+                msg = (
+                    'Index "{0}" age ({1}), period start: "{2}", period '
+                    'end, "{3}"'.format(
+                        index,
+                        age,
+                        start,
+                        end
+                    )
+                )
+                # Because time adds to epoch, smaller numbers are actually older
+                # timestamps.
+                inrange = ((age >= start) and (age <= end))
+                self.__excludify(inrange, exclude, index, msg)
+            except KeyError:
+                self.loggit.debug(
+                    'Index "{0}" does not meet provided criteria. '
+                    'Removing from list.'.format(index, source))
+                self.indices.remove(index)
 
     def iterate_filters(self, filter_dict):
         """